refactored add todo and todo id

+++ src/main/scala/objektwerks/Todo.scala
@@ -3,7 +3,7 @@ package objektwerks
 import scalikejdbc.{autoConstruct, ResultName, WrappedResultSet}
 import scalikejdbc.orm.CRUDMapper
 
-final case class Todo(id: Int = 0, task: String)
+final case class Todo(id: Long = 0, task: String)
 
 object Todo extends CRUDMapper[Todo]:
   override val tableName = "todo"
